requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
django==5.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
psycopg[binary]==3.2.1
    # via -r requirements.in
psycopg-binary==3.2.1
    # via psycopg
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dotenv==1.0.1
    # via -r requirements.in
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via psycopg
tzdata==2024.1
    # via
    #   django
    #   psycopg
